refactor: remove all backward-compatibility shims
Delete shim files that re-exported from canonical locations:

- app/graph_pipeline.py → app.pipeline.graph
- app/routing.py → app.pipeline.routing
- app/runners.py → app.pipeline.runners
- app/integrations/clients/ → app.services

Updated the 4 remaining importers to use canonical paths directly.

All 465 tests pass, lint clean, zero mypy errors.

Made-with: Cursor
