--- app/integrations/clients/__init__.py
@@ -1,53 +1,0 @@
-"""Backward-compatibility shim — canonical location is app.services."""
-
-from app.services import (
-    AWSBatchJobResult,
-    CoralogixClient,
-    GrafanaAccountConfig,
-    GrafanaClient,
-    HoneycombClient,
-    LogResult,
-    PipelineRunSummary,
-    PipelineSummary,
-    RootCauseResult,
-    S3CheckResult,
-    TracerClient,
-    TracerRunResult,
-    TracerTaskResult,
-    build_coralogix_logs_query,
-    get_grafana_client_from_credentials,
-    get_llm_for_reasoning,
-    get_llm_for_tools,
-    get_metric_statistics,
-    get_s3_client,
-    get_tracer_client,
-    get_tracer_web_client,
-    parse_root_cause,
-    reset_llm_singletons,
-)
-
-__all__ = [
-    "AWSBatchJobResult",
-    "CoralogixClient",
-    "GrafanaAccountConfig",
-    "GrafanaClient",
-    "HoneycombClient",
-    "LogResult",
-    "PipelineRunSummary",
-    "PipelineSummary",
-    "RootCauseResult",
-    "S3CheckResult",
-    "TracerClient",
-    "TracerRunResult",
-    "TracerTaskResult",
-    "build_coralogix_logs_query",
-    "get_grafana_client_from_credentials",
-    "get_llm_for_reasoning",
-    "get_llm_for_tools",
-    "get_metric_statistics",
-    "get_s3_client",
-    "get_tracer_client",
-    "get_tracer_web_client",
-    "parse_root_cause",
-    "reset_llm_singletons",
-]
