+++ pytest.ini
@@ -1,7 +1,7 @@
 [pytest]
 testpaths = tests app
 python_files = test_*.py *_test.py
-norecursedirs = tests/e2e cdk.out .git __pycache__ .pytest_cache .venv node_modules *.egg .tox
+norecursedirs = tests/e2e tests/deployment cdk.out .git __pycache__ .pytest_cache .venv node_modules *.egg .tox
 python_functions = test_*
 addopts = -v --tb=short
 filterwarnings =
