+++ tests/benchmarks/toolcall_model_benchmark/pipeline_benchmark.py
@@ -16,7 +16,7 @@ from typing import Any
 
 from langchain_core.runnables import RunnableConfig
 
-import app.graph_pipeline as graph_pipeline
+import app.pipeline.graph as graph_pipeline
 import app.services.llm_client as llm_mod
 from app.state import AgentState, make_initial_state
 from tests.benchmarks.toolcall_model_benchmark.pricing import estimate_run_cost_usd
